Find the kernel directly in memory rather than looking in its image on disk.
(This (ab)uses the undocumented ZwQuerySystemInformation function. But
since we're using it just to hook the kernel's import of another
undocumented function from the HAL...)

libpe no longer exists, and there is no longer a dependency on a
kernel image that matches the running one being present in system32/
at reboot time.